image-enhancement-system/requirements.txt
opencv-python>=4.5.0
numpy>=1.20.0
matplotlib>=3.4.0
scikit-image>=0.18.0
mysql-connector-python>=8.0.0
pillow>=8.0.0
pytest>=6.0.0
python-dotenv>=0.19.0
scikit-image>=0.18.0
scipy>=1.6.0
